adjust code style
